fix: properly fix clang warnings in photon source code
- Change IStream::preadv parameter from size_t to off_t, unifying with
  IFile::preadv and POSIX preadv semantics (-Woverloaded-virtual)
- Replace memmove with std::move_backward in headers.cpp
  (-Wnontrivial-memcall / -Wclass-memaccess)
- Add defaulted copy constructor to Timeout class (-Wdeprecated-copy)
- Add defaulted copy constructor/assignment to Object (-Wdeprecated-copy)

+++ net/http/message.cpp
@@ -161,7 +161,7 @@ ssize_t Message::readv(const struct iovec *iov, int iovcnt) {
     return m_body_stream->readv(iov, iovcnt);
 }
 
-ssize_t Message::preadv(const struct iovec *iov, int iovcnt, size_t off) {
+ssize_t Message::preadv(const struct iovec *iov, int iovcnt, off_t off) {
     if (!m_body_stream)
         LOG_ERROR_RETURN(EIO, -1, "body not readable");
     return m_body_stream->preadv(iov, iovcnt, off);
